refactor(web): move changes-panel banner and helper to a sibling file
