+++ apps/web/components/task/task-changes-panel.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { memo, useMemo, useCallback, createRef, useState, useEffect, useRef } from "react";
-import { IconAlertTriangle } from "@tabler/icons-react";
 import { PanelRoot, PanelBody } from "./panel-primitives";
+import { TruncatedFilesBanner, useSelectedFileKey } from "./changes-panel-banner";
 import { useToast } from "@/components/toast-provider";
 import { useAppStore } from "@/components/state-provider";
 import { useReviewSources, type ReviewSource } from "@/hooks/domains/session/use-review-sources";
@@ -532,39 +532,6 @@ const TaskChangesPanel = memo(function TaskChangesPanel({
     </PanelRoot>
   );
 });
-
-/** Composite key for the file currently selected in single-file mode. */
-function useSelectedFileKey(
-  mode: string,
-  filePath: string | undefined,
-  fileRepositoryName: string | undefined,
-): string | undefined {
-  return useMemo(
-    () =>
-      mode === "file" && filePath
-        ? reviewFileKey({ path: filePath, repository_name: fileRepositoryName })
-        : undefined,
-    [mode, filePath, fileRepositoryName],
-  );
-}
-
-/** Banner shown when the backend dropped files from a huge cumulative diff
- *  (large rebase) to keep the rendered row count bounded. */
-function TruncatedFilesBanner({ count }: { count: number }) {
-  if (count <= 0) return null;
-  return (
-    <div
-      data-testid="changes-truncated-banner"
-      className="flex items-center gap-2 px-4 py-2 text-xs text-yellow-600 bg-yellow-500/10 border-b border-yellow-500/20"
-    >
-      <IconAlertTriangle className="h-3.5 w-3.5 shrink-0" />
-      <span>
-        {count.toLocaleString()} more changed {count === 1 ? "file is" : "files are"} hidden — the
-        change set is too large to render in full.
-      </span>
-    </div>
-  );
-}
 
 function ChangesPanelContent({
   isLoading,
